Split test cases per flow
