+++ tests/test_oauth2_implicit_flow.py
@@ -1,88 +1,21 @@
 import time
-import multiprocessing
-import logging
-import urllib.request
 import re
 
 import requests
 import pytest
 from responses import RequestsMock
 
-
-from tests import authenticated_test_service
+from tests.oauth2_helper import (
+    authenticated_service,
+    token_cache,
+    TEST_SERVICE_HOST,
+    TIMEOUT,
+)
+from tests.auth_helper import get_header
 import requests_auth
 
 
-logger = logging.getLogger(__name__)
-
-
-TEST_SERVICE_PORT = 5001  # TODO Should use a method to retrieve a free port instead
-TEST_SERVICE_HOST = "http://localhost:{0}".format(TEST_SERVICE_PORT)
-TIMEOUT = 10
-
-
-def call(auth):
-    # Send a request to a dummy URL with authentication
-    requests.get("http://authorized_only", auth=auth)
-
-
-def get_header(responses, auth):
-    # Mock a dummy response
-    responses.add(responses.GET, "http://authorized_only")
-    # Send a request to this dummy URL with authentication
-    response = requests.get("http://authorized_only", auth=auth)
-    # Return headers received on this dummy URL
-    return response.request.headers
-
-
-def get_query_args(responses, auth):
-    # Mock a dummy response
-    responses.add(responses.GET, "http://authorized_only")
-    # Send a request to this dummy URL with authentication
-    response = requests.get("http://authorized_only", auth=auth)
-    # Return headers received on this dummy URL
-    return response.request.path_url
-
-
-def can_connect_to_server(port: int):
-    try:
-        response = urllib.request.urlopen(
-            "http://localhost:{0}/status".format(port), timeout=0.5
-        )
-        return response.code == 200
-    except:
-        return False
-
-
-def _wait_for_server_to_be_started(port: int):
-    for attempt in range(3):
-        if can_connect_to_server(port):
-            logger.info("Test server is started")
-            break
-        logger.info("Test server still not started...")
-    else:
-        raise Exception("Test server was not able to start.")
-
-
-@pytest.fixture(scope="module")
-def authenticated_service():
-    test_service_process = multiprocessing.Process(
-        target=authenticated_test_service.start_server, args=(TEST_SERVICE_PORT,)
-    )
-    test_service_process.start()
-    _wait_for_server_to_be_started(TEST_SERVICE_PORT)
-    yield test_service_process
-    test_service_process.terminate()
-    test_service_process.join(timeout=0.5)
-
-
-@pytest.fixture
-def token_cache():
-    yield requests_auth.OAuth2.token_cache
-    requests_auth.OAuth2.token_cache.clear()
-
-
-def test_oauth2_implicit_flow_url_is_mandatory(authenticated_service, token_cache):
+def test_oauth2_implicit_flow_url_is_mandatory():
     with pytest.raises(Exception) as exception_info:
         requests_auth.OAuth2Implicit(None)
     assert str(exception_info.value) == "Authorization URL is mandatory."
@@ -103,8 +36,6 @@ def test_oauth2_implicit_flow_token_is_not_reused_if_a_url_parameter_is_changing
 
     # Ensure that the new generated token will be different than previous one
     time.sleep(1)
-
-    logger.info("Requesting a custom token with a different parameter in URL.")
 
     auth2 = requests_auth.OAuth2Implicit(
         TEST_SERVICE_HOST
@@ -255,10 +186,11 @@ def test_oauth2_implicit_flow_post_failure_if_token_is_not_provided(
     authenticated_service, token_cache
 ):
     with pytest.raises(Exception) as exception_info:
-        call(
-            requests_auth.OAuth2Implicit(
+        requests.get(
+            "http://authorized_only",
+            auth=requests_auth.OAuth2Implicit(
                 TEST_SERVICE_HOST + "/do_not_provide_token", timeout=TIMEOUT
-            )
+            ),
         )
     assert str(exception_info.value) == "access_token not provided within {}."
 
@@ -267,11 +199,12 @@ def test_oauth2_implicit_flow_get_failure_if_token_is_not_provided(
     authenticated_service, token_cache
 ):
     with pytest.raises(Exception) as exception_info:
-        call(
-            requests_auth.OAuth2Implicit(
+        requests.get(
+            "http://authorized_only",
+            auth=requests_auth.OAuth2Implicit(
                 TEST_SERVICE_HOST + "/do_not_provide_token_as_anchor_token",
                 timeout=TIMEOUT,
-            )
+            ),
         )
     assert str(exception_info.value) == "access_token not provided within {}."
 
@@ -280,12 +213,13 @@ def test_oauth2_implicit_flow_post_failure_if_state_is_not_provided(
     authenticated_service, token_cache
 ):
     with pytest.raises(Exception) as exception_info:
-        call(
-            requests_auth.OAuth2Implicit(
+        requests.get(
+            "http://authorized_only",
+            auth=requests_auth.OAuth2Implicit(
                 TEST_SERVICE_HOST
                 + "/provide_token_as_access_token_but_without_providing_state",
                 timeout=TIMEOUT,
-            )
+            ),
         )
     assert re.match(
         "state not provided within {'access_token': \['.*'\]}.",
@@ -297,12 +231,13 @@ def test_oauth2_implicit_flow_get_failure_if_state_is_not_provided(
     authenticated_service, token_cache
 ):
     with pytest.raises(Exception) as exception_info:
-        call(
-            requests_auth.OAuth2Implicit(
+        requests.get(
+            "http://authorized_only",
+            auth=requests_auth.OAuth2Implicit(
                 TEST_SERVICE_HOST
                 + "/provide_token_as_anchor_access_token_but_without_providing_state",
                 timeout=TIMEOUT,
-            )
+            ),
         )
     assert re.match(
         "state not provided within {'access_token': \['.*'\]}.",
@@ -314,10 +249,11 @@ def test_oauth2_implicit_flow_failure_if_token_is_not_received_within_the_timeou
     authenticated_service, token_cache
 ):
     with pytest.raises(Exception) as exception_info:
-        call(
-            requests_auth.OAuth2Implicit(
+        requests.get(
+            "http://authorized_only",
+            auth=requests_auth.OAuth2Implicit(
                 TEST_SERVICE_HOST + "/do_not_redirect", timeout=TIMEOUT
-            )
+            ),
         )
     assert str(
         exception_info.value
@@ -349,188 +285,3 @@ def test_oauth2_implicit_flow_token_is_requested_again_if_expired(
     assert re.match("^Bearer .*", token2)
 
     assert token1 != token2
-
-
-def test_oauth2_authorization_code_flow_get_code_is_sent_in_authorization_header_by_default(
-    authenticated_service, token_cache, responses: RequestsMock
-):
-    auth = requests_auth.OAuth2AuthorizationCode(
-        TEST_SERVICE_HOST + "/provide_code_as_anchor_code",
-        "http://provide_access_token",
-        timeout=TIMEOUT,
-    )
-    responses.add(
-        responses.POST,
-        "http://provide_access_token",
-        json={
-            "access_token": "2YotnFZFEjr1zCsicMWpAA",
-            "token_type": "example",
-            "expires_in": 3600,
-            "refresh_token": "tGzv3JOkF0XG5Qx2TlKWIA",
-            "example_parameter": "example_value",
-        },
-    )
-    assert (
-        get_header(responses, auth).get("Authorization")
-        == "Bearer 2YotnFZFEjr1zCsicMWpAA"
-    )
-
-
-def test_oauth2_password_credentials_flow_token_is_sent_in_authorization_header_by_default(
-    authenticated_service, token_cache, responses: RequestsMock
-):
-    auth = requests_auth.OAuth2ResourceOwnerPasswordCredentials(
-        "http://provide_access_token",
-        username="test_user",
-        password="test_pwd",
-        timeout=TIMEOUT,
-    )
-    responses.add(
-        responses.POST,
-        "http://provide_access_token",
-        json={
-            "access_token": "2YotnFZFEjr1zCsicMWpAA",
-            "token_type": "example",
-            "expires_in": 3600,
-            "refresh_token": "tGzv3JOkF0XG5Qx2TlKWIA",
-            "example_parameter": "example_value",
-        },
-    )
-    assert (
-        get_header(responses, auth).get("Authorization")
-        == "Bearer 2YotnFZFEjr1zCsicMWpAA"
-    )
-
-
-def test_oauth2_client_credentials_flow_token_is_sent_in_authorization_header_by_default(
-    authenticated_service, token_cache, responses: RequestsMock
-):
-    auth = requests_auth.OAuth2ClientCredentials(
-        "http://provide_access_token",
-        username="test_user",
-        password="test_pwd",
-        timeout=TIMEOUT,
-    )
-    responses.add(
-        responses.POST,
-        "http://provide_access_token",
-        json={
-            "access_token": "2YotnFZFEjr1zCsicMWpAA",
-            "token_type": "example",
-            "expires_in": 3600,
-            "refresh_token": "tGzv3JOkF0XG5Qx2TlKWIA",
-            "example_parameter": "example_value",
-        },
-    )
-    assert (
-        get_header(responses, auth).get("Authorization")
-        == "Bearer 2YotnFZFEjr1zCsicMWpAA"
-    )
-
-
-def test_okta_client_credentials_flow_token_is_sent_in_authorization_header_by_default(
-    authenticated_service, token_cache, responses: RequestsMock
-):
-    auth = requests_auth.OktaClientCredentials(
-        "test_okta", client_id="test_user", client_secret="test_pwd", timeout=TIMEOUT
-    )
-    responses.add(
-        responses.POST,
-        "https://test_okta/oauth2/default/v1/token",
-        json={
-            "access_token": "2YotnFZFEjr1zCsicMWpAA",
-            "token_type": "example",
-            "expires_in": 3600,
-            "refresh_token": "tGzv3JOkF0XG5Qx2TlKWIA",
-            "example_parameter": "example_value",
-        },
-    )
-    assert (
-        get_header(responses, auth).get("Authorization")
-        == "Bearer 2YotnFZFEjr1zCsicMWpAA"
-    )
-
-
-def test_okta_client_credentials_flow_token_is_sent_in_authorization_header_by_default_using_helper(
-    authenticated_service, token_cache, responses: RequestsMock
-):
-    auth = requests_auth.okta(
-        requests_auth.OAuth2Flow.ClientCredentials,
-        "test_okta",
-        client_id="test_user",
-        client_secret="test_pwd",
-        timeout=TIMEOUT,
-    )
-    responses.add(
-        responses.POST,
-        "https://test_okta/oauth2/default/v1/token",
-        json={
-            "access_token": "2YotnFZFEjr1zCsicMWpAA",
-            "token_type": "example",
-            "expires_in": 3600,
-            "refresh_token": "tGzv3JOkF0XG5Qx2TlKWIA",
-            "example_parameter": "example_value",
-        },
-    )
-    assert (
-        get_header(responses, auth).get("Authorization")
-        == "Bearer 2YotnFZFEjr1zCsicMWpAA"
-    )
-
-
-def test_header_api_key_requires_an_api_key():
-    with pytest.raises(Exception) as exception_info:
-        requests_auth.HeaderApiKey(None)
-    assert str(exception_info.value) == "API Key is mandatory."
-
-
-def test_query_api_key_requires_an_api_key():
-    with pytest.raises(Exception) as exception_info:
-        requests_auth.QueryApiKey(None)
-    assert str(exception_info.value) == "API Key is mandatory."
-
-
-def test_header_api_key_is_sent_in_X_Api_Key_by_default(responses: RequestsMock):
-    auth = requests_auth.HeaderApiKey("my_provided_api_key")
-    assert get_header(responses, auth).get("X-Api-Key") == "my_provided_api_key"
-
-
-def test_query_api_key_is_sent_in_api_key_by_default(responses: RequestsMock):
-    auth = requests_auth.QueryApiKey("my_provided_api_key")
-    assert get_query_args(responses, auth) == "/?api_key=my_provided_api_key"
-
-
-def test_header_api_key_can_be_sent_in_a_custom_field_name(responses: RequestsMock):
-    auth = requests_auth.HeaderApiKey("my_provided_api_key", "X-API-HEADER-KEY")
-    assert get_header(responses, auth).get("X-Api-Header-Key") == "my_provided_api_key"
-
-
-def test_query_api_key_can_be_sent_in_a_custom_field_name(responses: RequestsMock):
-    auth = requests_auth.QueryApiKey("my_provided_api_key", "X-API-QUERY-KEY")
-    assert get_query_args(responses, auth) == "/?X-API-QUERY-KEY=my_provided_api_key"
-
-
-def test_basic_authentication_send_authorization_header(responses: RequestsMock):
-    auth = requests_auth.Basic("test_user", "test_pwd")
-    assert (
-        get_header(responses, auth).get("Authorization")
-        == "Basic dGVzdF91c2VyOnRlc3RfcHdk"
-    )
-
-
-def test_basic_and_api_key_authentication_can_be_combined(responses: RequestsMock):
-    basic_auth = requests_auth.Basic("test_user", "test_pwd")
-    api_key_auth = requests_auth.HeaderApiKey("my_provided_api_key")
-    header = get_header(responses, basic_auth + api_key_auth)
-    assert header.get("Authorization") == "Basic dGVzdF91c2VyOnRlc3RfcHdk"
-    assert header.get("X-Api-Key") == "my_provided_api_key"
-
-
-def test_basic_and_api_key_authentication_can_be_combined_deprecated(
-    responses: RequestsMock,
-):
-    basic_auth = requests_auth.Basic("test_user", "test_pwd")
-    api_key_auth = requests_auth.HeaderApiKey("my_provided_api_key")
-    header = get_header(responses, requests_auth.Auths(basic_auth, api_key_auth))
-    assert header.get("Authorization") == "Basic dGVzdF91c2VyOnRlc3RfcHdk"
-    assert header.get("X-Api-Key") == "my_provided_api_key"
